+++ xncml/core.py
@@ -1,12 +1,20 @@
 from collections import OrderedDict
 from pathlib import Path
+from typing import Any
 from warnings import warn
 
 import xmltodict
 
 
 class Dataset(object):
-    """This is a class for reading and manipulating NcML file"""
+    """This is a class for reading and manipulating NcML file.
+
+    Note that NcML documents are used for two distinct purposes:
+      - an XML description of NetCDF structure and metadata;
+      - create virtual NetCDF datasets, e.g. an aggregation of multiple files.
+
+    This class supports both types of uses.
+    """
 
     def __init__(self, filepath: str = None, location: str = None):
         """
@@ -25,8 +33,8 @@ class Dataset(object):
             # Convert all dictionaries to lists of dicts to simplify the internal logic.
             self.ncroot = xmltodict.parse(
                 self.filepath.read_text(),
-                process_namespaces=True,
                 force_list=['variable', 'attribute', 'group', 'dimension'],
+                process_namespaces=True,
                 namespaces={
                     'http://www.unidata.ucar.edu/namespaces/netcdf/ncml-2.2': None,
                     'https://www.unidata.ucar.edu/namespaces/netcdf/ncml-2.2': None,
@@ -278,3 +286,160 @@ class Dataset(object):
         xml_output = xmltodict.unparse(self.ncroot, pretty=True)
         with open(path, 'w') as fd:
             fd.write(xml_output)
+
+    def to_cf_dict(self):
+        """Convert internal representation to a CF-JSON dictionary.
+
+        The CF-JSON specification includes `data` for variables, but if the data is not within the NcML,
+        it cannot be included in the JSON representation.
+
+        Returns
+        -------
+        Dictionary with `dimensions` and `variables` keys. May also optionally include an `attributes` key and a
+        `groups` key. Additional keys prefixed with `@` may be included for <netcdf> tag attributes,
+        for example `@location`.
+
+        References
+        ----------
+        http://cf-json.org/specification
+        """
+        res = OrderedDict()
+        nc = self.ncroot['netcdf']
+
+        for key, val in nc.items():
+            if key[0] == '@':
+                res[key] = val
+            if key == 'dimension':
+                res.update(_dims_to_json(val))
+            if key == 'group':
+                res.update(_groups_to_json(val))
+            if key == 'attribute':
+                res.update(_attributes_to_json(val))
+            if key == 'variable':
+                res.update(_variables_to_json(val))
+
+        return res
+
+
+def _dims_to_json(dims: list) -> dict:
+    """The dimensions object has dimension id:size as its key:value members."""
+    out = OrderedDict()
+    for dim in dims:
+        if int(dim['@length']) > 1:
+            out[dim['@name']] = int(dim['@length'])
+
+    return {'dimensions': out}
+
+
+def _groups_to_json(groups: list) -> dict:
+    out = OrderedDict()
+    for group in groups:
+        name = group['@name']
+        out[name] = OrderedDict()
+        if 'attribute' in group:
+            out[name].update(_attributes_to_json(group['attribute']))
+        if 'group' in group:
+            out[name].update(_groups_to_json(group['group']))
+
+    return {'groups': out}
+
+
+def _attributes_to_json(attrs: list) -> dict:
+    """The attributes object contains arbitrary attributes as its key:value members."""
+    out = OrderedDict()
+    for attr in attrs:
+        try:
+            out[attr['@name']] = _cast(attr)
+        except ValueError as exc:
+            warn(f"Could not cast {attr['@name']}:\n{exc}")
+
+    return {'attributes': out}
+
+
+def _variables_to_json(variables: list) -> dict:
+    """The variables definition object has variable id:object as its key:value members.
+
+    Each variable object MUST include shape, attributes and data objects.
+    The shape field is an array of dimension IDs which correspond to the array ordering of the variable data.
+    """
+    out = OrderedDict()
+
+    # Put coordinate variables first
+    for var in variables:
+        if _is_coordinate(var):
+            out[var['@name']] = None
+
+    for var in variables:
+        name = var['@name']
+        out[name] = OrderedDict()
+
+        if '@shape' in var:
+            out[name]['shape'] = var['@shape'].split(' ')
+
+        if '@type' in var:
+            out[name]['type'] = var['@type']
+
+        if 'attribute' in var:
+            out[name].update(_attributes_to_json(var['attribute']))
+
+        if 'values' in var:
+            out[name]['data'] = _cast(var)
+
+    return {'variables': out}
+
+
+def _cast(obj: dict) -> Any:
+    """Cast attribute value to the appropriate type."""
+    from xncml.parser import DataType, nctype
+
+    value = obj.get('@value') or obj.get('values')
+    typ = DataType(obj.get('@type', 'String'))
+    if value is not None:
+        if isinstance(value, str):
+            if typ in [DataType.STRING, DataType.STRING_1]:
+                return value
+
+            sep = ' '
+            values = value.split(sep)
+            return list(map(nctype(typ), values))
+        elif isinstance(value, dict):
+            raise NotImplementedError(obj)
+        else:
+            return value
+
+
+def _is_coordinate(var):
+    """Return True is variable is a coordinate."""
+
+    # Variable is 1D and has same name as dimension
+    if var.get('@shape', '').split(' ') == [var['@name']]:
+        return True
+
+    lat_units = ['degrees_north', 'degreeN', 'degree_N', 'degree_north', 'degreesN', 'degrees_N']
+    lon_units = ['degrees_east', 'degreeE', 'degree_E', 'degree_east', 'degreesE', 'degrees_E']
+    names = [
+        'latitude',
+        'longitude',
+        'time',
+        'air_pressure',
+        'altitude',
+        'depth',
+        'geopotential_height',
+        'height',
+        'height_above_geopotential_datum',
+        'height_above_mean_sea_level',
+        'height_above_reference_ellipsoid',
+    ]
+
+    if 'attribute' in var:
+        attrs = _attributes_to_json(var['attribute'])
+
+        # Check units
+        if attrs.get('units', '') in lon_units + lat_units:
+            return True
+
+        # Check long_name and standard_name
+        if attrs.get('long_name', attrs.get('standard_name', '')) in names:
+            return True
+
+    return False