+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Implement `Dataset.rename_dataset_attribute`. By @huard
 - Allow empty `Dataset` creation. By @huard
 - Add support in `Dataset` for NcML documents using the `ncml` namespace. By @huard
+- Implement `Dataset.to_cf_dict` method to export CF-JSON dictionary. By @huard.
 
 
 0.1 Initial release (2022-11-24)